add back events

+++ lib/global/global.h
@@ -26,7 +26,7 @@ struct GlobalVars
         int16_t sensors2;
         int16_t knk_rtd;
 
-        uint8_t vetable[16 * 16];
+        float vetable[16 * 16];
         uint16_t rpm_table[16];
         int16_t map_table[16];
     } ms;
